Fixes snippet rendering bug. Debug tool.
Fun little debug tool has no texture or name, but when right clicked on
a block it will add the id and meta to the chat, and if the block is a
tile entity it will also add ALL the declared fields and values. Not
sure how useful it will actually be, just thought it would be fun.