More Gradle and API work.
A few more tweaks to build.gradle, including sperating the configs into
gradle.properties.

Along with that, I'm increasing the Forge requirement. We were getting a
little behind. It's now the latest: 10.12.1.1085

This change didn't seem to have any noticable effects within Core.

Worked on IDismantleableTile a bit. It's not done and requires lots of
testing, but any changes shouldn't break anything else, so I'm going to
commit it "as-is".