Equivalency Map and Set, SimpleItemHelper.
The new Equivalency API is an attempt to create a Map and Set
implementation that allow asymmetric equality. That is, A == B but B !=
A. Allowing 'A' to encapsulate 'B' and act as a wrapper while 'B'
remains specific. For example, quad.equivalency(rectangle, square) ==
true but quad.equivalency(square, rectangle) == false.

SimpleItemHelper is ment to replace ItemRegUtil and ItemStackUtil, with
a better interface based API, with support for json SimpleItems. More
work needs to be done but preliminary testing with Geocraft branch shows
that the principal works.